requirements.txt
pillow==10.3.0
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-socketio[asgi]==5.11.2
python-multipart==0.0.9
